+++ egui_demo_lib/src/lib.rs
@@ -12,6 +12,7 @@
     clippy::checked_conversions,
     clippy::dbg_macro,
     clippy::debug_assert_with_mut_call,
+    clippy::disallowed_method,
     clippy::doc_markdown,
     clippy::empty_enum,
     clippy::enum_glob_use,
@@ -21,12 +22,17 @@
     clippy::explicit_into_iter_loop,
     clippy::fallible_impl_from,
     clippy::filter_map_next,
+    clippy::flat_map_option,
     clippy::float_cmp_const,
     clippy::fn_params_excessive_bools,
+    clippy::from_iter_instead_of_collect,
     clippy::if_let_mutex,
+    clippy::implicit_clone,
     clippy::imprecise_flops,
     clippy::inefficient_to_string,
     clippy::invalid_upcast_comparisons,
+    clippy::large_digit_groups,
+    clippy::large_stack_arrays,
     clippy::large_types_passed_by_value,
     clippy::let_unit_value,
     clippy::linkedlist,
@@ -35,8 +41,10 @@
     clippy::manual_ok_or,
     clippy::map_err_ignore,
     clippy::map_flatten,
+    clippy::map_unwrap_or,
     clippy::match_on_vec_items,
     clippy::match_same_arms,
+    clippy::match_wild_err_arm,
     clippy::match_wildcard_for_single_variants,
     clippy::mem_forget,
     clippy::mismatched_target_os,
@@ -46,6 +54,7 @@
     clippy::mutex_integer,
     clippy::needless_borrow,
     clippy::needless_continue,
+    clippy::needless_for_each,
     clippy::needless_pass_by_value,
     clippy::option_option,
     clippy::path_buf_push_overwrite,
@@ -53,6 +62,8 @@
     clippy::ref_option_ref,
     clippy::rest_pat_in_fully_bound_structs,
     clippy::same_functions_in_if_condition,
+    clippy::semicolon_if_nothing_returned,
+    clippy::single_match_else,
     clippy::string_add_assign,
     clippy::string_add,
     clippy::string_lit_as_bytes,
@@ -153,4 +164,20 @@ fn test_egui_zero_window_size() {
         let clipped_meshes = ctx.tessellate(shapes);
         assert!(clipped_meshes.is_empty(), "There should be nothing to show");
     }
+}
+
+// ----------------------------------------------------------------------------
+
+/// Time of day as seconds since midnight. Used for clock in demo app.
+pub(crate) fn seconds_since_midnight() -> Option<f64> {
+    #[cfg(feature = "chrono")]
+    {
+        use chrono::Timelike;
+        let time = chrono::Local::now().time();
+        let seconds_since_midnight =
+            time.num_seconds_from_midnight() as f64 + 1e-9 * (time.nanosecond() as f64);
+        Some(seconds_since_midnight)
+    }
+    #[cfg(not(feature = "chrono"))]
+    None
 }